Rename QueueRepository to QueueEntryRepository for clarity
