+++ ShopkeepersQuiz.Api/Services/Questions/QuestionService.cs
@@ -12,7 +12,7 @@ using ShopkeepersQuiz.Api.Models.Configuration;
 using ShopkeepersQuiz.Api.Models.Questions;
 using ShopkeepersQuiz.Api.Models.Queues;
 using ShopkeepersQuiz.Api.Repositories.Questions;
-using ShopkeepersQuiz.Api.Repositories.Queues;
+using ShopkeepersQuiz.Api.Repositories.QueueEntries;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -25,18 +25,18 @@ namespace ShopkeepersQuiz.Api.Services.Questions
 		const int TimeDriftBufferSeconds = 1;
 
 		private readonly IQuestionRepository _questionRepository;
-		private readonly IQueueRepository _queueRepository;
+		private readonly IQueueEntryRepository _queueEntryRepository;
 		private readonly QuestionSettings _questionSettings;
 		private readonly IMemoryCache _cache;
 
 		public QuestionService(
 			IQuestionRepository questionRepository,
-			IQueueRepository queueRepository,
+			IQueueEntryRepository queueRepository,
 			IOptions<QuestionSettings> questionSettings,
 			IMemoryCache cache)
 		{
 			_questionRepository = questionRepository;
-			_queueRepository = queueRepository;
+			_queueEntryRepository = queueRepository;
 			_questionSettings = questionSettings.Value;
 			_cache = cache;
 		}
@@ -51,7 +51,7 @@ namespace ShopkeepersQuiz.Api.Services.Questions
 				return questionQueue;
 			}
 
-			questionQueue = (await _queueRepository.GetUpcomingQueueEntries()).ToList();
+			questionQueue = (await _queueEntryRepository.GetUpcomingQueueEntries()).ToList();
 			if (questionQueue.Count >= questionCount)
 			{
 				_cache.Set(CacheKeys.QuestionQueue, JsonConvert.SerializeObject(questionQueue));
@@ -205,7 +205,7 @@ namespace ShopkeepersQuiz.Api.Services.Questions
 				throw new ArgumentNullException(nameof(questionQueue));
 			}
 
-			IEnumerable<QueueEntry> existingQueueEntries = await _queueRepository.GetUpcomingQueueEntries();
+			IEnumerable<QueueEntry> existingQueueEntries = await _queueEntryRepository.GetUpcomingQueueEntries();
 
 			IEnumerable<QueueEntry> queueEntriesToAdd = questionQueue
 				.Where(x => x.StartTimeUtc >= DateTime.UtcNow)
@@ -213,7 +213,7 @@ namespace ShopkeepersQuiz.Api.Services.Questions
 
 			foreach (var entry in queueEntriesToAdd)
 			{
-				await _queueRepository.CreateQueueEntry(entry);
+				await _queueEntryRepository.CreateQueueEntry(entry);
 			}
 		}
 	}
